fix: compiler warnings

+++ src/domain/game.cpp
@@ -1,6 +1,7 @@
 #include "game.hpp"
 #include "domain/event.hpp"
 #include "hand.hpp"
+#include <cstddef>
 #include <stdexcept>
 
 namespace {
@@ -40,7 +41,7 @@ namespace domain {
             deal_card_to_dealer();
         }
 
-        for (int i = 0; i < player_.hand_count(); ++i) {
+        for (size_t i = 0; i < player_.hand_count(); ++i) {
             evaluate_hand(i);
         }
     }
@@ -108,7 +109,7 @@ namespace domain {
         deal_card_to_player_hand(new_hand_idx);
     }
 
-    void Game::evaluate_hand(int hand_index) {
+    void Game::evaluate_hand(size_t hand_index) {
         const auto& player_hand = player_.hands()[hand_index].hand;
         if (player_hand.is_busted()) {
             player_.set_outcome_for_hand(HandOutcome::Busted, hand_index);
